scsi: ufs: sprd: add UMS9230 UFS support
Add support for the UFS controller found in the Unisoc UMS9230 SoC to
the existing driver for UMS9620. This controller is not very similar to
the one used in UMS9620, but it also requires setting various syscon
registers for initialization. The initialization sequence is based on
code from Unisoc's downstream Android kernel.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>