dt-bindings: iommu: sprd: add sprd,iova-range property
The Unisoc IOMMU has a configurable virtual address range, which should
not be too large because the IOMMU only supports a flat page table. It
is sometimes necessary to ensure that certain addresses fall within this
range, e.g. when creating a pass-through mapping for a boot framebuffer,
or to configure different range sizes for different IOMMUs. Define a new
property to allow setting the address range.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>